Refactored UserContext logic to break into smaller/lighter individual modules
