+++ apps/public-reference/pages/_app.jsx
@@ -2,7 +2,7 @@ import React from "react"
 import App from "next/app"
 import Router from "next/router"
 import "@bloom-housing/ui-components/styles/index.scss"
-import { addTranslation, UserProvider } from "@bloom-housing/ui-components"
+import { addTranslation, UserProvider, ConfigProvider } from "@bloom-housing/ui-components"
 import { headScript, bodyTopTag, pageChangeHandler } from "../src/customScripts"
 import { AppSubmissionContext, blankApplication } from "../lib/AppSubmissionContext"
 import { loadApplicationFromAutosave } from "../lib/ApplicationConductor"
@@ -94,9 +94,11 @@ class MyApp extends App {
           syncApplication: this.syncApplication,
         }}
       >
-        <UserProvider apiUrl={process.env.listingServiceUrl}>
-          <Component {...pageProps} />
-        </UserProvider>
+        <ConfigProvider apiUrl={process.env.listingServiceUrl}>
+          <UserProvider>
+            <Component {...pageProps} />
+          </UserProvider>
+        </ConfigProvider>
       </AppSubmissionContext.Provider>
     )
   }
