+++ shared/ui-components/src/authentication/index.ts
@@ -1,0 +1,2 @@
+export { UserContext, UserProvider } from "./UserContext"
+export { useAuthenticatedClient } from "./useAuthenticatedClient"
