+++ shared/ui-components/src/config/index.ts
@@ -1,0 +1,1 @@
+export { ConfigProvider, ConfigContext } from "./ConfigContext"
